refactored triggers samples into snippet format
